misc lint fixes, pass 1

+++ x/stakeibc/keeper/claim_test.go
@@ -98,7 +98,7 @@ func (s *KeeperTestSuite) SetupClaimUndelegatedTokens() ClaimUndelegatedTestCase
 				Amount:      redemptionAmount,
 			}},
 			ICAAccountType: types.ICAAccountType_REDEMPTION,
-			Timeout:        uint64(types.DefaultICATimeoutNanos),
+			Timeout:        types.DefaultICATimeoutNanos,
 		},
 	}
 }
